api/requirements.txt
# Repo root api/ for Vercel Python serverless (Vercel uses FastAPI app natively)
fastapi>=0.121.2
openai>=1.0.0
python-dotenv>=1.0.0
